+++ core/shared/src/main/scala/zio/config/experimental/IndexedFlat.scala
@@ -1,7 +1,6 @@
-package zio.config
+package zio.config.experimental
 
 import zio._
-// import zio.ConfigProvider.Flat
 
 trait IndexedFlat extends ConfigProvider.Flat {
   import IndexedFlat._
